Attempt at auto-pan based on tutorial
Tutorial found at:
https://developers.arcgis.com/labs/android/display-and-track-your-location/

In attempting to do this, I converted Java (a language I don't really
know) to Kotlin (a language that I don't really know). Unsurprisingly
the finished product is nonfunctional: the application boots to a blank
screen.

I tried to use the logging that I found in AndroidStudio but didn't
turn up much. I am putting the Kotlin version of this project aside
and will re-attempt it in Java, which is what all of these tutorials
are actually written in.